pyproject.toml
[project]
name = "project-atlas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "click>=8.1.0",
    "fastapi>=0.115.0",
    "langchain>=0.3.27",
    "langchain-ibm>=0.3.18",
    "langchain-openai>=0.3.35",
    "langchain-ollama>=1.0.0",
    "langgraph>=0.6.10",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=13.5.2,<13.6",
    "semgrep>=1.139.0",
    "fastmcp>=1.0",
    "uvicorn>=0.32.0",
    "vt-py>=0.18.0",
    "reportlab>=4.0.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "jupyterlab>=4.4.10",
    "pre-commit>=4.0.1",
    "pylint>=3.3.9",
    "pytest>=8.4.2",
]

[tool.black]
line-length = 100

[project.scripts]
project-atlas = "project_atlas.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
